Created an apps directory and vendor directory.
Added a base.html file to the main templates folder.

Moved the original app "geoapi" into an apps folder for better clarity on the layout of the project.

Created a vendor directory for apps that have to be added and modified manually. First app added is a registration app and the files were modified to work with the project.

Inside the main directory updated the settings.py file for the new apps and added a new url for the registration.